+++ modules/main/src/main/java/com/nisovin/shopkeepers/config/migration/ConfigMigrations.java
@@ -20,7 +20,8 @@ public class ConfigMigrations {
 			new ConfigMigration3(),
 			new ConfigMigration4(),
 			new ConfigMigration5(),
-			new ConfigMigration6()
+			new ConfigMigration6(),
+			new ConfigMigration7()
 	);
 
 	public static int getLatestVersion() {